unified db seeders, fixed parsers

+++ NotiFIITBot/Database/Models/NotificationConfig.cs
@@ -10,7 +10,7 @@ public class UserNotificationConfig
     public User User { get; set; }
 
     [Column("lesson_id")]
-    public int LessonId { get; set; }
+    public long LessonId { get; set; }
     public LessonModel Lesson { get; set; }
